+++ lib/Controller/MetadataController.php
@@ -2,6 +2,7 @@
 namespace OCA\Metadata\Controller;
 
 use OC\Files\Filesystem;
+use OCA\Metadata\AppInfo\Application;
 use OCA\Metadata\Service\MetadataService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\JSONResponse;
@@ -37,7 +38,7 @@ class MetadataController extends Controller {
 				return new JSONResponse(
 					array(
 						'response' => 'error',
-						'msg' => $this->language->t('No metadata found.')
+						'msg' => Application::getL10N()->t('No metadata found.')
 					)
 				);
 			}