+++ lib/Service/TiffParser.php
@@ -1,7 +1,7 @@
 <?php
 namespace OCA\Metadata\Service;
 
-class TiffParser {
+class TiffParser extends FileReader {
     const BYTE = 1;
     const ASCII = 2;
     const SHORT = 3;
@@ -20,7 +20,7 @@ class TiffParser {
 
         if (($data === "II\x2A\x00") || ($data === "MM\x00\x2A")) {     // ID
             $intel = ($data[0] === 'I');
-            $ifdOffs = $this->readInt($hnd, $intel);
+            $ifdOffs = self::readInt($hnd, $intel);
 
             $this->parseTiffIfd($hnd, $pos, $intel, $ifdOffs);
         }
@@ -29,12 +29,12 @@ class TiffParser {
     public function parseTiffIfd($hnd, $pos, $intel, $ifdOffs) {
         while (!feof($hnd) && ($ifdOffs !== 0)) {
             fseek($hnd, $pos + $ifdOffs);               // Go to IFD
-            $tagCnt = $this->readShort($hnd, $intel);
+            $tagCnt = self::readShort($hnd, $intel);
 
             for ($i = 0; $i < $tagCnt; $i++) {
-                $tagId = $this->readShort($hnd, $intel);
-                $tagType = $this->readShort($hnd, $intel);
-                $count = $this->readInt($hnd, $intel);
+                $tagId = self::readShort($hnd, $intel);
+                $tagType = self::readShort($hnd, $intel);
+                $count = self::readInt($hnd, $intel);
                 $offsetOrData = fread($hnd, 4);
                 $size = -1;
                 switch ($tagType) {
@@ -58,7 +58,7 @@ class TiffParser {
                         if ($size <= 4) {
                             $offsetOrData = substr($offsetOrData, 0, $size);
                             if ($count === 1) {
-                                $offsetOrData = $this->unpackShort($intel, $offsetOrData);
+                                $offsetOrData = self::unpackShort($intel, $offsetOrData);
                             } else {
                                 $f = $intel? 'v' : 'n';
                                 $d = unpack($f.'a/'.$f.'b', $offsetOrData);
@@ -70,7 +70,7 @@ class TiffParser {
                     case self::SLONG:
                         $size = $count * 4;
                         if ($size <= 4) {
-                            $offsetOrData = $this->unpackInt($intel, $offsetOrData);
+                            $offsetOrData = self::unpackInt($intel, $offsetOrData);
                         }
                         break;
                     case self::RATIONAL:
@@ -79,7 +79,7 @@ class TiffParser {
                         break;
                 }
                 if ($size > 4) {
-                    $offsetOrData = $this->unpackInt($intel, $offsetOrData);
+                    $offsetOrData = self::unpackInt($intel, $offsetOrData);
                 }
 
                 if ($size > 0) {
@@ -94,7 +94,7 @@ class TiffParser {
                 }
             }
 
-            $ifdOffs = $this->readInt($hnd, $intel);
+            $ifdOffs = self::readInt($hnd, $intel);
             if (($ifdOffs !== 0) && ($ifdOffs < ftell($hnd))) {         // Never go back
                 $ifdOffs = 0;
             }
@@ -105,25 +105,5 @@ class TiffParser {
 
     protected function processTag($hnd, $pos, $intel, $tagId, $tagType, $count, $size, $offsetOrData) {
         return null;
-    }
-
-    protected function readShort($hnd, $intel) {
-        return $this->unpackShort($intel, fread($hnd, 2));
-    }
-
-    protected function readInt($hnd, $intel) {
-        return $this->unpackInt($intel, fread($hnd, 4));
-    }
-
-    protected function readRat($hnd, $intel) {
-        return $this->readInt($hnd, $intel) . '/' . $this->readInt($hnd, $intel);
-    }
-
-    protected function unpackShort($intel, $data) {
-        return unpack(($intel? 'v' : 'n').'d', $data)['d'];
-    }
-
-    protected function unpackInt($intel, $data) {
-        return unpack(($intel? 'V' : 'N').'d', $data)['d'];
     }
 }